requirements.txt
django-bcrypt==0.9.2
django-celery==3.1.17
django-ipware==1.1.3
django-redis==4.3.0
django-webpack-loader==0.3.0
djangorestframework==3.3.3
celery=3.1.23
django-crispy-forms==1.6.0
